Updated database

+++ MovieTime/entities/src/main/java/il/cshaifasweng/OCSFMediatorExample/entities/Movie.java
@@ -57,7 +57,7 @@ public class Movie implements  Serializable
 		this.launchDate = launchDate;
 	}
     
-	public Movie(String name, String imageSrc, String duration, Integer popular, String genre) {
+	public Movie(String name, String imageSrc, String duration, Double popular, String genre) {
 
 		super();
 		Name = name;
